Use the node/core naming convention  - Add _core file to hold the implementation  - _node file now holds the executable  - Modify CMakeLists to point to the new executable
